slurm-pred/requirements.txt
PyYAML==6.0.2
simple-slurm==0.3.6
